+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+## Unreleased
+- Adds [`MediumType.prototype.name`][] as a way to get the full type as
+  a string.
+
+  ```javascript
+  new MediumType("application/rdf+xml").name // application/rdf+xml
+  ```
+
+[`MediumType.prototype.name`]: https://github.com/moll/js-medium-type/blob/master/doc/API.md#mediumType.name
+
 ## 0.1.337 (Nov 4, 2014)
 - First ultra spiritual release.